create a new tool CountMutationTypes
The new tool gets an VCF file as an input and create a GATK report with the percentages of each mutation type (e.g. A->G, A->T...).
It allow the user to filter sites that will be count based of JXEL or based on the varait quals
A user can aslo print 12 VCF files (one for each mutation) with the VCF line of the mutations that were counted.