Fixed MLEAC and QUAL inaccuracy in GeneralPloidyExactAFCalculator.
The problem whas that the MLE table calculation aborted "unlikely"
genotype combinations to aggresively.

This also uncovered another bug where GeneralPloidyExactAFCalculation
makes a slightly different use of StateTracker
as compared to DiploidExactAFCalculation. We have changed StateTracker
generalizing it to be able to work with both using code behaviors.

Story:
-----

  * https://www.pivotaltracker.com/story/show/78920568

Changes:
-------

  * Fixes in GeneralPloidyExactAFCalculator.
  * Needed changes in StateTracker API and its consequences in DiploidExactAFCalculation.
  * Updated affected integrated tests' MD5s after fixing the GeneralPloidyExactAF.