Small optimizations to the joint calling code.
Thanks to profiling I noticed that the determineCoefficient() method was being called too often.
Because it returns a constant result in half of the invocations, its value should be cached when possible.
Also, the various calls to getLog10Likelihoods() showed up in the profiler, so I pulled those out too.

All told, it speeds up the genotyping by about 10 percent according to the profiler.